LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.rect.nonfinite.worker-expected.txt


PASS rect() with Infinity/NaN is ignored
